Add wildcard certificate support
Preliminary support for issuing wildcard certificates using the dns-01
challenge. The feature is gated behind a config flag, allowWildcards.

The relevant sections of the standard are as follows:

RFC8555 §7.1.3:
> Any identifier of type "dns" in a newOrder request MAY have a
> wildcard domain name as its value.  A wildcard domain name consists
> of a single asterisk character followed by a single full stop
> character ("*.") followed by a domain name as defined for use in the
> Subject Alternate Name Extension by [RFC5280].  An authorization
> returned by the server for a wildcard domain name identifier MUST NOT
> include the asterisk and full stop ("*.") prefix in the authorization
> identifier value.  The returned authorization MUST include the
> optional "wildcard" field, with a value of true.

RFC8555 §7.1.4:
> Wildcard domain names (with "*" as the first
> label) MUST NOT be included in authorization objects.  If an
> authorization object conveys authorization for the base domain of a
> newOrder DNS identifier containing a wildcard domain name, then the
> optional authorizations "wildcard" field MUST be present with a value
> of true.

What this attempts to say is that in Authorizations the wildcard prefix
needs to be stripped from identifiers, but in Orders they must stay.

This is confirmed by analyzing the objects sent by Boulder in response
to certbot -d '*.example.com,example.com', reveals:

* the order object contains identifiers *.example.com and example.com
  (this means, no deduplication is needed when parsing requested
  identifiers in the Order)
* the authorization object for *.example.com shows identifier
  example.com, with an additional wildcard:true attribute, and only the
  dns-01 challenge being allowed.
* there is a separate authorization object for example.com with the
  usual parameters (all challenges, no wildcard attribute).

Having this difference handled by the JSON serializer allows us to
continue using a single Identifier object for both Order and Authz.

Note that the use of str.removeprefix() has bumped the minimum Python
version to 3.9.